Added customizable number of clique island (FourCliqueGraph).

+++ src/main/java/org/sample/POD.java
@@ -2,5 +2,5 @@ package org.sample;
 
 public class POD
 {
-    public static long nodesNum = 903;
+    public static int nodesNum = 903;
 }
